examples/sample-app/requirements.txt
flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
google-cloud-sql-connector[pg8000]==1.13.0
pg8000==1.31.2
requests==2.32.3
